Fix up prop drilling
