+++ frontend/src/scenes/notebooks/Notebook/Editor.tsx
@@ -1,6 +1,6 @@
 import posthog from 'posthog-js'
-import { useActions } from 'kea'
-import { useCallback, useRef } from 'react'
+import { useActions, useValues } from 'kea'
+import { useCallback, useMemo, useRef } from 'react'
 
 import { Editor as TTEditor } from '@tiptap/core'
 import { EditorContent, useEditor } from '@tiptap/react'
@@ -25,32 +25,31 @@ import { lemonToast } from '@posthog/lemon-ui'
 import { NotebookNodeType } from '~/types'
 import { NotebookNodeImage } from '../Nodes/NotebookNodeImage'
 
-import { EditorFocusPosition, EditorRange, JSONContent, Node, NotebookEditor, textContent } from './utils'
+import { EditorFocusPosition, EditorRange, JSONContent, Node, textContent } from './utils'
 import { SlashCommandsExtension } from './SlashCommands'
 import { BacklinkCommandsExtension } from './BacklinkCommands'
 import { NotebookNodeEarlyAccessFeature } from '../Nodes/NotebookNodeEarlyAccessFeature'
 import { NotebookNodeSurvey } from '../Nodes/NotebookNodeSurvey'
 import { InlineMenu } from './InlineMenu'
 import NodeGapInsertionExtension from './Extensions/NodeGapInsertion'
+import { notebookLogic } from './notebookLogic'
+import { sampleOne } from 'lib/utils'
 
 const CustomDocument = ExtensionDocument.extend({
     content: 'heading block*',
 })
 
-export function Editor({
-    onCreate,
-    onUpdate,
-    onSelectionUpdate,
-    placeholder,
-}: {
-    onCreate: (editor: NotebookEditor) => void
-    onUpdate: () => void
-    onSelectionUpdate: () => void
-    placeholder: ({ node }: { node: any }) => string
-}): JSX.Element {
+const PLACEHOLDER_TITLES = ['Release notes', 'Product roadmap', 'Meeting notes', 'Bug analysis']
+
+export function Editor(): JSX.Element {
     const editorRef = useRef<TTEditor>()
-    const logic = insertionSuggestionsLogic()
-    const { resetSuggestions, setPreviousNode } = useActions(logic)
+
+    const { shortId } = useValues(notebookLogic)
+    const { setEditor, onEditorUpdate, onEditorSelectionUpdate } = useActions(notebookLogic)
+
+    const { resetSuggestions, setPreviousNode } = useActions(insertionSuggestionsLogic)
+
+    const headingPlaceholder = useMemo(() => sampleOne(PLACEHOLDER_TITLES), [shortId])
 
     const updatePreviousNode = useCallback(() => {
         const editor = editorRef.current
@@ -67,7 +66,17 @@ export function Editor({
                 gapcursor: false,
             }),
             ExtensionPlaceholder.configure({
-                placeholder: placeholder,
+                placeholder: ({ node }: { node: any }) => {
+                    if (node.type.name === 'heading' && node.attrs.level === 1) {
+                        return `Untitled - maybe.. "${headingPlaceholder}"`
+                    }
+
+                    if (node.type.name === 'heading') {
+                        return `Heading ${node.attrs.level}`
+                    }
+
+                    return ''
+                },
             }),
             FloatingMenu.extend({
                 onSelectionUpdate() {
@@ -181,7 +190,7 @@ export function Editor({
         onCreate: ({ editor }) => {
             editorRef.current = editor
 
-            onCreate({
+            setEditor({
                 getJSON: () => editor.getJSON(),
                 getText: () => textContent(editor.state.doc),
                 getEndPosition: () => editor.state.doc.content.size,
@@ -225,8 +234,8 @@ export function Editor({
                 },
             })
         },
-        onUpdate: onUpdate,
-        onSelectionUpdate: onSelectionUpdate,
+        onUpdate: onEditorUpdate,
+        onSelectionUpdate: onEditorSelectionUpdate,
     })
 
     return (
